+++ src/backend/storage/shmem.h
@@ -15,7 +15,7 @@
  * in every process.  Shared memory "pointers" are actually
  * offsets relative to the start of the shared memory region(s).
  */
-typedef unsigned int SHMEM_OFFSET;
+typedef unsigned long SHMEM_OFFSET;
 #define INVALID_OFFSET (-1)
 #define BAD_LOCATION (-1)
 
@@ -27,14 +27,14 @@ extern SHMEM_OFFSET ShmemBase;
 
 /* coerce an offset into a pointer in this process's address space */
 #define MAKE_PTR(xx_offs)\
-  (ShmemBase+((unsigned int)(xx_offs)))
+  (ShmemBase+((unsigned long)(xx_offs)))
 
 /* coerce a pointer into a shmem offset */
 #define MAKE_OFFSET(xx_ptr)\
-  (SHMEM_OFFSET) (((unsigned int)(xx_ptr))-ShmemBase)
+  (SHMEM_OFFSET) (((unsigned long)(xx_ptr))-ShmemBase)
 
 #define SHM_PTR_VALID(xx_ptr)\
-  (((unsigned int)xx_ptr) > ShmemBase)
+  (((unsigned long)xx_ptr) > ShmemBase)
 
 /* cannot have an offset to ShmemFreeStart (offset 0) */
 #define SHM_OFFSET_VALID(xx_offs)\
@@ -47,8 +47,8 @@ typedef struct SHM_QUEUE {
 } SHM_QUEUE;
 
 /* shmem.c */
-int *ShmemAlloc();
-int *ShmemInitStruct();
+long *ShmemAlloc();
+long *ShmemInitStruct();
 /* dont declare this so we avoid nested include files */
 /*HTAB *ShmemInitHash(); */
 
@@ -68,8 +68,8 @@ typedef struct {
         /* string name */
   char          key[BTABLE_KEYSIZE];
         /* location in shared mem */
-  unsigned int  location;
+  unsigned long  location;
         /* numbytes allocated for the structure */
-  unsigned int  size;
+  unsigned long  size;
 } BindingEnt;
 #endif	/* !defined(ShMemIncluded) */
